+++ C4DT/src/net/arctics/clonk/parser/c4script/ast/StringLiteral.java
@@ -141,7 +141,7 @@ public final class StringLiteral extends Literal<String> {
 	public void reportErrors(C4ScriptParser parser) throws ParsingException {
 		
 		// warn about overly long strings
-		int max = parser.getContainer().getIndex().getEngine().getCurrentSettings().maxStringLen;
+		long max = parser.getContainer().getIndex().getEngine().getCurrentSettings().maxStringLen;
 		if (max != 0 && getLiteral().length() > max) {
 			parser.warningWithCode(ParserErrorCode.StringTooLong, this, getLiteral().length(), max);
 		}
